Teach the divider code how to handle the reverse attribute.
When a color has been reversed, then it is being pulled from the field
opposite what we expect.  The foreground is stored in the background
attribute and vice versa.  This commit causes the divider to check the
reversed properties of the segment color to decide how to properly setup
the divider colors.